+++ api/index.js
@@ -13,6 +13,7 @@ import cors from "cors";
 import multer from "multer";
 import searchRoutes from "./routes/searches.js";
 import announcementRoutes from "./routes/announcements.js";
+import videoRoutes from "./routes/videos.js";
 
 // middlewares
 app.use((req, res, next) => {
@@ -55,6 +56,7 @@ app.use("/api/likes", likeRoutes);
 app.use("/api/relationships", relationshipRoutes);
 app.use("/api/searches", searchRoutes);
 app.use("/api/announcements", announcementRoutes);
+app.use("/api/videos", videoRoutes);
 
 // our user id would be like this = currentUser.id and userInfo.id